Move generic logic from TabArenaContext into AbstractArenaContext
AbstractArenaContext now owns everything that depends only on the
context's method/task metadata: artifact loading (load_raw, load_repo,
generate_repo, load_*_results, load_configs_hyperparameters,
load_config_results_multi), HPO/portfolio simulation (run_hpo,
combine_hpo, generate_*_trajectories, simulate_portfolio*,
run_portfolio*), plotting (plot_tuning_trajectories*,
plot_runtime_per_method, generate_per_dataset_tables,
generate_all_figs), find_missing, and fillna_metrics.

The base is no longer abstract: the two preset hooks are concrete and
reject all names, so the class is directly instantiable with an
explicit methods list + TaskMetadataCollection. TabArenaContext keeps
only what is TabArena-specific: the v0.1 presets, the extra subset
predicates/default subsets, RF-default fillna/calibration defaults,
and evaluate_all (hardcoded paper baselines).

The custom-datasets example now uses AbstractArenaContext directly and
no longer depends on TabArenaContext.
